An attempt to fix some panics in the tests.
We have reports of panics in the tests, which I think are the result of
items being raced to be pushed in the channel after the `workCh` is already closed.
We had those in the scale tests.

Even if that isn't the case for the tests (i.e. more bugs are in the
pool), I think this way it's more reliable.

Question: should Go return error, or just swallow the fact that new work
was not scheduled.