Hollow out our ConfigMaps. (#3047)
This change hollows out the ConfigMaps that we use with Knative
in order to walk a very fine line that enables us to deploy
knative/serving via `ko` or `kubectl apply` while still preserving
edits made to the serving ConfigMaps.  This fine line is dancing
around the way the `kubectl apply` annotation
`kubectl.kubernetes.io/last-applied-configuration` is used to update
resources.

Essentially, this annotation is attached with the body of the current
apply, and subsequent applications take it upon themselves to remove
keys that were present in the prior apply, but not in the current apply.

By reserving all of our ConfigMap keys to be specified by the user
via `kubectl edit cm -nknative-serving config-foo`, which does not
update the annotation, we ensure that subsequent applies won't remove
the keys because those keys were never a part of the previous apply.

One of the key scenarios this unlocks is the ability to preserve
users' domain settings in `config-domain.yaml` across applies.

Fixes: https://github.com/knative/serving/issues/2668