Allow VolumeMounts to be specified in Container. (#3061)
* Allow VolumeMounts to be specified in Container.

This extends the knative/serving API to allow for ConfigMaps and Secrets to be mounted as volumes.  All other forms of volume are disallowed.

The container's volume mounts must include all of the provided volumes, they must be ReadOnly (this is defaulted in the webhook), and only a MountPath may be specified.

These are then passed through to the Deployment, and the rest is up to K8s.

This adds two conformance tests, one for each of ConfigMap and Secret, which simply expects the Service to echo back the randomized text read from the mounted volume.

Related: https://github.com/knative/serving/issues/1504

* Fix the type name used in the spec changed.

* Incorporate new runtime contract wording.

Add validation checks for the mountPath of volumeMounts (for conflicts with reserved paths).

Improve resource naming in the conformance testing.

* Rename functional options and improve test errors.

* Check absolute paths, reserved directory parents, and a few other pieces of feedback.