Add conformance test to validate sys parameters are read-only (#3101)
* Add conformance test to validate sys parameters are read-only

This adds a conformance test to validate the SHOULD clause for sysctls
in the runtime contract. This ensures that sysctls mounted in the
container are read-only. It does this by looking at the mount options
for the filesystem in /proc/mounts. This change also breaks out the
runtime image handler to prevent it from becoming too large.

* Address PR comments

* Change optional to omitempty