Take sleep out of activator tests. (#3195)
* Fix flaky activator tests.

These tests used a sleep to allow requests to run concurrently and assumed that the sleep is high enough. That caused flakiness.

This swaps these sleeps out with unbuffered channels to be able to accurately (and deterministically) controll the concurrent access to a critical point in the code to trigger the condition under test.

* Some more documentation.

* Simplify with var.

* Pull values into constants.

* Add timeout handling so tests have better output than timing out.