Use a fake recorder in the SyncHandler tests (#3269)
We are seeing a large number of tests complaining about `t.Log` being called after the test has completed, and the callstack shows that this is coming through the event broadcaster stuff, which is passed a reference to the logger.

The default event broadcaster stuff is bypassed if passed a recorder in the reconciler options, so pass in a fake instead to try and reduce flakes.

Fixes: https://github.com/knative/serving/issues/3209