Allow for the format of Routes of a Service to be configurable
By default Routes are set to `ServiceName.Namespace.Domain`.
This change allows for someone to set a `RouteTemplate` field in the
`config-network` configMap with a golang template expression so they
can define any other URL pattern to use.

E.g. `{{.Name}}-{{.Namespace}}.{{.Domain}}` will use `-` instead of `.`
For us this is useful because we have an environment where the TLS cert
is wildcarded for `*.domain` and the wildcard only allows for one level
above the domain. Obviously, people need to be aware of what format they
choose because it could lead to restrictions. E.g., in this example, the use
of `-` in the Namespace could be problematic.  But this their choice.

As of now they can substitute Name, Namespace or Domain. We could
choose to add other values later if we wish by just adding them to the
data structure.

One thing we might want to consider is allowing the Route to be a URL
so that people can choose something other than just DNS names, for example:
`Domain/Namespace/Name` might be a valid choice for some environments.

Docs and test are coming but I wanted to get this out there for discussion
before I went too far in case the entire idea is rejected.

Signed-off-by: Doug Davis <dug@us.ibm.com>