Remove migration of deprecated generation labels and annotation
Addresses #643

In 0.4, Revisions had generation labels migrated toward use of
`/configurationGeneration` label to determine a Configuration's
latest created Revision.

This commit leaves the deprecated `/configurationMetadataGeneration`
label intact to allow rollback to 0.4, but removes the migration
of the deprecated label, along with the deletion of an annotation
previously used for the same purpose.

Functional dependency on `/configurationMetadataGeneration` is dropped
in #3325.