Remove several pieces of code we no longer need. (#3360)
* Remove several pieces of code we no longer need.

When I previously introduced the network probing in the activator it replaced the need for the "request rewinder" and the "retrying transport", which have been inactive since it went in.

This also removes the handler in the activator that enforces a limit on the request's content-length, which was introduced in #1292 (alongside the Rewinder) to bound the size of the request we're willing to buffer into memory.  Now that we don't "rewind" the request, we no longer have cause to limit the payload size.

Fixes: https://github.com/knative/serving/issues/3359

* Also remove the FilteringHandler.

This handler was used to remove "count" headers from the request when they cycled back to the activator.

We now no longer need the count headers, so this has been completely removed.

* Rework the way we measure attempts to include probes, soften logging.