Use a value for empty strings in stats reporter (#3358)
Some apps do not define service names (i.e., service name is ""),
and stats reporter takes it as missing service names. The error is
reported as "Failed to export to Prometheus: inconsistent label
cardinality" in logs.

This fix uses a special value instead of "" so that stats reporter has
enough dimensions.