add test for websocket connection through activator (#3343)
* add test for websocket connection through activator

Addresses #3254

Adds a test that stands up a service based on the wsserver
test image, polls it until it is scaled down to zero replicas,
then initiates a websocket connection.

* use helper function to build deployment name

* clean up WaitForScaleToZero

allows WaitForScaletoZero to accept the deploymentName directly
as an argument. Looks up the ScaleToZeroGrace period from the
autoscaler instead of hardcoding a value.

* fix comment

* remove empty line

* check error from WaitForScaleToZero in grpc tests

* extract gracePeriodMultiplier constant, capitalise log message

* build deploymentName directly

* fix comment

* pull grace period multiplier out of scaleToZeroGracePeriod function