Update the breaker test. (#3401)
Main change is top replace the mutex that had an empty wait sequence
with a channel (alternative could be a WaitGroup()).

Also some superficial changes.

I am still not completely satisfied, since the test violates the
sanctity of breaker in the `concurrentRequest` method and panics
in the waitForQueue -- both of which I'd like to change at some later
point.