Add a new stat metric in queue to prevent double counting
Currently a request pending in both the activator and queue proxy is
counted twice, once in the activator, and one in the queue proxy. This
change fixes it by letting queue report a concurrency metric for proxied
requests and letting the autoscaler discount such concurrency when it
calculates the total concurrency for scaling decision.