Add URL to TrafficTarget Status
This change adds URL to the output of each named Traffic Target. The URL
is prefixed with a scheme. Currently we only support subdomains for
named traffic targets and 'http' as the scheme. However, this change enables
clients to grab the URL from the API rather than making assumptions on
scheme and URL format which paves the way for operator customization.

Fixes #3450