Disallow unknown fields (again)
This adds back the logic in the webhook that disallows unknown fields.

We had disabled this because of versioning, but the idea of silently accepting unknown fields is an undesireable failure mode.

We also can't currently detect / reject typos and such because of this, which continues to be a real hit to usability.

Fixes: https://github.com/knative/serving/issues/3309