Add a new stat metric in queue to prevent double counting (#3477)
* Add a new stat metric in queue to prevent double counting

Currently a request pending in both the activator and queue proxy is
counted twice, once in the activator, and one in the queue proxy. This
change fixes it by letting queue report a concurrency metric for proxied
requests and letting the autoscaler discount such concurrency when it
calculates the total concurrency for scaling decision.

* Update test to stay compatible with a recent change.

* Incorporate feedback.

* Minor changes to address feedback on naming, logging, and documentation.

* Improved readability with comments and a little refactoring.