This implements "BYO Revision name".
With this change, in any context where a user might specify a
`RevisionTemplateSpec`, they may now provide their own Revision name:

```yaml
kind: Configuration
metadata:
  name: config
spec:
  revisionTemplate:
    metadata:
      name: config-foo # Must have config- prefix.
    spec:
      ...
```

The intent of this is to enable GitOps scenarios which cannot be covered
well by our generated names today.  With this in the context of Service,
a new Revision may be staged via something like:

```yaml
kind: Service
metadata:
  name: svc
spec:
  release:
    revisions: [svc-2019-03-27, svc-2019-03-28]
    percent: 1
    metadata:
      name: svc-2019-03-28
    spec:
      ...
```

In the above example, once `svc-2019-03-28` is `Ready` it will
automatically begin receiving 1% of traffic.

The restrictions on BYO name are as follows:
 * The name must change (or be removed) any time the body of the
  `revisionTemplate` changes.
 * The name must be prefixed by the name of the enclosing Service
  or Configuration with a trailing `-`.
 * The name must be unique across the namespace (the above is intended
  to help with this, and preserve some of the self-documenting nature
  of the prefix, but it doesn't guarantee collision avoidance).