Return 'BadRequest' if a probe request doesn't match.
Arguably, returning 'ServiceUnavailable' if a probe response does not match the expected target isn't semantically the right error response. Rather, the request sent in doesn't match the expectation of the probe handler, thus 'BadRequest' makes more sense here.

This has the side effect of not triggering any istio-internal retries for these requests, making the retry-interval-policy entirely up to the prober.