Relax the collision validation to allow spec equivalence.
This change is inspired by a use case that arose in a slack
conversation around `git revert` with GitOps.

The example is the following is committed:

```yaml
spec:
  template:
    metadata:
      name: mattmoor-foo
    spec:
      A
```

Then the following update is committed:

```yaml
spec:
  template:
    metadata:
      name: mattmoor-bar
    spec:
      B
```

Then a problem is discovered with the above and a `git revert`
is done to the prior state:

```yaml
spec:
  template:
    metadata:
      name: mattmoor-foo
    spec:
      A
```

In the original model, this case would be rejected by the Service
controller due to the generation mismatch between the existing
`mattmoor-foo` and the latest generation of the underlying
Configuration.