Remove an imperative use of Conditions.
Today we key off of a strange mix of CreationTimestamp and LastTransitionTime
on Ready as a "clock" for deployment progress.  I believe that this is a
throwback to a world where Ready changes based on the active state.

During deployment, this clock is unreliable because the conditions vary a fair
amount.  Once Ready, we shouldn't see this particular condition again.

This also switches to using the CreationTimestamp on the Endpoints instead of
the Revision, since the Endpoints no longer come/go.

Related: https://github.com/knative/pkg/issues/357