Add reconciliation for the protocol type. (#3661)
* Add reconciliation for the protocol type.

Currently we have lots of resources with PT unset, and this upsets
the downstream systems, since the defaulting is done in the revision
code. And we do not want to propagate that logic to the downstream
objects (otherwise we could just avoided the field in the first place).

So this adds temporary reconciliation for the time being, and in 0,7 it
can be removed.

* address review