Improve the autoscale test. (#3690)
* Improve the autoscale test.

First, it started as debugging the presubmit failure. And I find myself
scrolling through pages and pages and pages of basically meaningless
lines of "request took %v" lines.
So, decided I, we should obliterate that spam.
But then I noticed other stuff, and then some more and I had to fix it
all.
And this is how this PR came to be.

* request is modified in spoof