Resolve tags to digest more generically
Previously, when resolving a tag to a digest, we'd fetch the reference
explicitly as an image. This is fine in the general case, but would only
work for amd64/linux images. By using remote.Get instead, we fetch the
reference with a broader set of Accept headers, allowing schema 1 and
manifest lists as well.

Since we don't resolve manifest lists to images, we'll pass through
multi-platform images to the nodes, which partially unblocks support for
other platforms.