Change a few fields to pointers and use accessors to fetch them. (#3722)
This change abstracts how we access the `corev1.Container` within `RevisionSpec` and the `RevisionTemplateSpec` within `ConfigurationSpec`, now both are accessed through getters `GetContainer()` and `GetTemplate()` respectively.  This is in preparation for the "lemonade" proposal to introduce the v1beta1 api in parallel to the v1alpha1 API, but is itself harmless (even if we don't go that way).

This also makes `TimeoutSeconds` a pointer, so we can distinguish not-specified from `0`.

I made this change by renaming the two fields (and then reverting it) to make sure I got everything.