Fix the misleading comment in the table.go (#3733)
The comment says the objects are "Sets" but they are cetainly not.
In fact they are ordered lists and matched index by index with
the actual create/update/delete/etc objects encountered during the
reconciliation (FWIW, the order is actually backwards).

I can be open to fix this later to actually work like a set, since
currently say we expect A B C to be updated, but the updates were A B D
C, we'll report D as a mismatch and then an unexpected update of C,
while in reality the only mismatch is action D.

Also produce comments for actions.go.