Drop the ConcurrencyModel fields from PodAutoscaler. (#3738)
As far as I can tell this has virtually always been deprecated since its
inception, and the one place that creates these also updates them (Victor
added this in 0.5 to backfill fields we need to expand its responsibilities).

We also have validation that not both fields are specified, but default
from one field to the other (guaranteeing both are set).

At a minimum as of 0.5 there should be zero instances of this field.