Simplify the ClusterIngress creation logic.
Instead of treating active / inactive completely differently, we now treat
them largely the same; however, instead of using the revision service we use
the activator service.  The main side-effect of this is that instead of
getting a single split for all inactive revisions, we end up with one per
inactive revision.

This also makes the logic to append a Revision header name unconditional, we
now always append a revision name header with the precedence:
1. Inactive Revision with the largest non-zero percent.
1. Revision with the largest non-zero percent, if we have no inactive splits.

Both of these are necessary changes for Revision-managed activation anyways,
and we should be appending the revision headers per-subroute.