This implements "BYO Revision name". (#3562)
* This implements "BYO Revision name".

With this change, in any context where a user might specify a
`RevisionTemplateSpec`, they may now provide their own Revision name:

```yaml
kind: Configuration
metadata:
  name: config
spec:
  revisionTemplate:
    metadata:
      name: config-foo # Must have config- prefix.
    spec:
      ...
```

The intent of this is to enable GitOps scenarios which cannot be covered
well by our generated names today.  With this in the context of Service,
a new Revision may be staged via something like:

```yaml
kind: Service
metadata:
  name: svc
spec:
  release:
    revisions: [svc-2019-03-27, svc-2019-03-28]
    percent: 1
    metadata:
      name: svc-2019-03-28
    spec:
      ...
```

In the above example, once `svc-2019-03-28` is `Ready` it will
automatically begin receiving 1% of traffic.

The restrictions on BYO name are as follows:
 * The name must change (or be removed) any time the body of the
  `revisionTemplate` changes.
 * The name must be prefixed by the name of the enclosing Service
  or Configuration with a trailing `-`.
 * The name must be unique across the namespace (the above is intended
  to help with this, and preserve some of the self-documenting nature
  of the prefix, but it doesn't guarantee collision avoidance).

* Relax the collision validation to allow spec equivalence.

This change is inspired by a use case that arose in a slack
conversation around `git revert` with GitOps.

The example is the following is committed:

```yaml
spec:
  template:
    metadata:
      name: mattmoor-foo
    spec:
      A
```

Then the following update is committed:

```yaml
spec:
  template:
    metadata:
      name: mattmoor-bar
    spec:
      B
```

Then a problem is discovered with the above and a `git revert`
is done to the prior state:

```yaml
spec:
  template:
    metadata:
      name: mattmoor-foo
    spec:
      A
```

In the original model, this case would be rejected by the Service
controller due to the generation mismatch between the existing
`mattmoor-foo` and the latest generation of the underlying
Configuration.

* Move away from abusing apis.Immutable for BYO name validation.

This replaces the current abuse to `apis.Immutable` for checking
certain properties of BYO Revision name in the webhook.  It replaces
them with the context-based approach we are moving towards to more
holistically replace `apis.Immutable` with `apis.Validate`.

* Incorporate review feedback from @vagababov

Fix some lint issues.

Moves to the shared context stuff in knative/pkg

Fix cmp.Diff orientation.

* Disallow BYO name in conjunction with generateName in the parent.

* Conflicting revision names should be fatal.

* SafeDiff to ShortDiff