Do not scale when there is no metric, and some cleanups. (#3771)
Currently when there is no metrics (desiredScale == -1), we reset
desiredScale to minScale, causing pods mistakenly scaled down to
minScale.  Also, when the autoscaler is restarted, it may have
PodAutoscaler.Status unknown and when the pod is scaled down to 0, it
will be bounced back to 1.

Revamp the scaler.Scale() logic to fix these issues, and also to make it
easier to read.

Some minor cleanups to make the code match the intent (e.g.,
isPAOwnedByRevision() returns true when the PA is owned by the revision,
rather than the other way around).