Fix the problem where old PAs go into unknown state.
Origianally (even before my changes) the old updatePAStatus
method would change the status of the PA to unknown, even if it's
inactive. The reason for that is that after restart for an inactive
revision the desired scale is always -1.
Now, this was not much of a problem in the previous life, since when
activated the desired scale would become 1 and everything would unfold
as desired. But in this new brave world, this causes SKS to reconcile to
serve more, which removes the activator endpoints from the public k8s
service and prohibits our future endeavours like positive handoff.

The change is simple: basically ignore want==-1 scale events, when
changing status. This works because:
if want == -1 and we're inactive => stay inactive
if want == -1 and we're active => stay active
if want == -1 and we're activating => continue activating.