Shift the e2e tests to use the v1beta1 subset. (#3853)
The most superficial part of thst is shifting the fields we use in tests from
`revisionTemplate` to `template`, `container` to `containers`, `name` to `tag`.

Things get more interesting when we look at Service "modes", where I've changed
things as follows:
 * "run latest" tests now test the path where we omit and default `traffic`
 * "release" tests not test an inline `traffic` block.
 * "manual" tests have been reworked to deal with Service and `traffic`.

The remaining "legacy" testing paths are:
 * We still test that the build integration works, but due to our contextual
   validation, it is only allowed within the legacy modes.  These tests also
   cover `runLatest`, `revisionTemplate` and `container`.
 * Upgrade testing still functions in terms of fields that existed at 0.5.

So while the bulk of our coverage has shifted to the new fields, we do have
basic coverage that the old way still works.

In order to have the tests consume `tag`, this also modifies the Route
controller to surface both `tag` and `name` in status, and `latestRevision`.