Various cleanups and bug fixes. (#3876)
We have accumulated some bugs and crud in our code, so this does a pass
over the code and cleans up some stuff.

The most important is the removal of the copy of Transport in the
transports.go. Transport is not a copyable type since it has mutex and
other shareds state inside. So replace it with a brand new transport.

Rest is just minor stuff.