Externalize creation of alternative test namespace. (#3878)
`TestMultipleNamespace` creates a new namespace from the test itself. That namespace has an unpredictable name. Some environments (Openshift for instance) need additional properties attached to a namespace to successfully launch and run new Knative services.

This externalizes the creation of an "alternative" namespace to be used by tests that want to test cross-namespace compatibility. This namespace follows the same lifecycle as the original `serving-tests` namespace.