Introduce a PodScalable duck type for use with Knative autoscaling.
The traditional model for autoscaling reconciliation involves using the
scale client on each reconcile to establish the current desired/observed
replica counts for the ScaleTargetRef.  Unlike much of the rest of our
system, this involves client-based reads from the API server, which are
not cached by an informer.  The `/scale` interface also gives us a
limited set of information:
 * Desired Scale
 * Observed Scale
 * Selectors

This PR replaces the use of the `/scale` client in the KPA with a new
`PodScalable` duck type, which abstracts the rough shape of the
Kubernetes resources that support `/scale` currently.

Fetching the current state of the `PodScalable` is done via a
`duck.InformerFactory` that returns `PodScalable` resources and can take
advantage of an underlying informer cache (so nop Reconciles shouldn't
cross the network).

Changing the desired replica count of the `PodScalable` is done via
a patch through the "dynamic" client.