Move deployment logic in to deployment dir
Move the remaining logic from under the revision controller in to the
new deployment dir. This has the benefit of route controller no longer
needing to import revision controller for serviceport constants and
allows other systems to refer to deployment creation logic. This also
plublic-ifys makeLabels and makeSelectors as
deployment.MakeRevision{Labels,Selectors}.