Make queue port listen only on port.
Currently we configure k8s service and the deployment to expose
only one port.
But the queue-proxy binary itself actually listens on two ports,
wasting the port number and the cycles to maintain the port.
Let's not do that but rather tell queue-proxy what port it should
listen on, since revision knows that already.