Scale down a revision when it is crashlooping.
Currently a revision can fail and the pod autoscaler keeps trying to
activate with no luck, leading to crashlooping.  The fix is to scale
down the pods when their revision failed.

As a byproduct, fix a bug that updates an immutable field of
PodAutoscaler.Spec, making the reconcilation of Revision's Active
condition deadcode.