Export TAG to fix our annotation manipulation. (#3995)
`generate-yamls.sh` performs a sed to rewrite annotations when the `TAG` bash variable is set, which it is in `hack/release.sh` after flag processing.  However, because this variable wasn't exported the child process didn't inherit it and the sed wasn't being performed.  This simply exports TAG before calling `generate-yaml.sh`.

Fixes: https://github.com/knative/serving/issues/3626