Reduce the idle timeouts in the default transports. (#3996)
This helps to deal with the problem we're seeing
with the lean Istio, wherein the connection to the K8s
service is not terminated when the underlying endpoint changes,
since in the absence of the sidecar while the socket is alive,
nothing disturbs the connection.
This stops us from scaling to 0.