Drop unneeded extra goroutine in the autoscaler.
The scaleChannel design seems to be dated back to an older autoscaler design. In the current implementation, the blocking behavior of `tickScaler` is exactly the same, be it decoupled via an extra scaleChannel or not.

This simplifies the autoscaler implementation to only use one goroutine driven by a ticker, inlining the once "decoupled" goroutine.