Attempt to fix our probes with mTLS enabled.
Currently, we check for the kubelet's user-agent to attempt to distinguish
between proper requests and probes, however, with Istio mTLS enabled they
rewrite probes, but they do not spoof the Kubelet's User-Agent.  This leads
to all manner of failure modes (see linked issue for one, another is likely
that the probe traffic will prevent scaling to zero).

To combat this, in the places we currently key off of the User-Agent we will
now be sensitive to a header, which we will explicitly pass in our Activator
probes, and rewrite user probes to provide.

Fixes: https://github.com/knative/serving/issues/3751