Don't overwrite K-Original-Host
When we're clearing the Host header, only put it into K-Original-Host when
it isn't already specified.  This enables us to front Knative services
with an Istio VirtualService for "vanity domains" that rewrite from e.g.
`www.mattmoor.io` to `foo.bar.mattmoor.io` by sticking the original host
in this header.

Fixes: https://github.com/knative/serving/issues/4029