Only list services within the same namespace. (#4069)
In the cleanup logic for the "Kevin" work we list the K8s services for a route
by the label selector with the route name, but we mistakenly did this cluster-wide.

This means that we'd discover services in other namespaces, and then (very likely)
fail deleting the service with that name in our namespace, leading the route to
never become ready.

This could happen if I have Routes in parallel namespaces with the same name, but
different "tags".  However, @tcnghia and I saw this playing with @bbrowning's
Ambassador implementation for ClusterIngress, which creates a K8s service in the
ambassador namespace as part of the ClusterIngress implementation.

Related to: https://github.com/knative/serving/issues/3419