Introduce ReadyPodCounter.UpdateName()
Currently, when Autoscaler.Update() is called, the code creates a
new ReadyPodCounter. This means that the implementation of the
ReadyPodCounter is effectively fixed to a particular concrete
type, in this case the endpointAddressCounter.

Introducing the new method makes it possible to update the value
of the name in-place, so that the original concrete type injected
through New() is retained.