Remove UpdateableReadyPodCounter type
In previous commits, the UpdateableReadyPodCounter type was introduced
to ensure that, upon autoscaler.Update(), the correct service name was
set on the UpdateableReadyPodCounter.

In conversation on PR knative/serving#4140, @vagabov pointed out that
the service name is intended to be immutable[0]. This means that
updating the service name is not a scenario that needs to be accounted
for.

[0] https://github.com/knative/serving/pull/4140#discussion_r286516170