Rename NewUnscopedEndpointsCounter
NewUnscopedEndpointsCounter -> NewFixedEndpointsListCounter

Normally that would be the headline, but it's too long for
Github's narrow requirement.

The rename reflects the "unscoped" was meant to be symmetrical
with NewScopedEndpointsCounter, which takes an endpoints
informer, namespace and name as parameters.

But actually, this ReadyPodsCounter does not do an "unscoped"
version of the "scoped" ReadyPodsCounter -- that is, it does
not use an informer to retrieve all endpoints across all
namespaces. Instead it takes a predetermined list of endpoints
that are collected by the calling code.