Output info message when clusteringress's status was updated successfully (#4144)
When `config-domain` configmap was updated, event log from
clusteringress sometimes tells us that `Failed to update status for
ClusterIngress`. Although this error would be solved automatically by
next reconcile loop, we don't know whether it was solved or not
because there are no success log.

To improve it, this patch adds success logs when clusteringress
was updated successfully.