Make pod counters an injectable type in autoscalers (#4140)
* Introduce ReadyPodCounter type

* Use ReadyPodCounter, not FetchReadAddressCount

This is a preliminary step towards pulling this counter into a
variable on the struct.

* Make ReadyPodCounter at autoscaler create & update

* Inject ReadyPodCounter into autoscaler at creation

* Introduce ReadyPodCounter.UpdateName()

Currently, when Autoscaler.Update() is called, the code creates a
new ReadyPodCounter. This means that the implementation of the
ReadyPodCounter is effectively fixed to a particular concrete
type, in this case the endpointAddressCounter.

Introducing the new method makes it possible to update the value
of the name in-place, so that the original concrete type injected
through New() is retained.

* Rm endpointsInformer from autoscaler.New() params

Previously, this reference was necessary in order to use
resources.FetchReadyAddressCount(). But now a ReadyPodCounter
is injected at creation time, which carries its own reference
to the endpointsLister (not the endpointsInformer as such).

* Replace FetchReadyAddressCount() with ReadyCount()

- Make FetchReadyAddressCount() private
- Remove test of private fetchReadyAddressCount() function

The logic previously exercised by TestFetchReadyAddressCount
is covered by TestEndpointAddressCounter.

* Replace ReadyAddressCount w/ unscoped ReadyCount

The goal of this commit was to make the ReadyAddressCount()
method private, so that the resources package consistently
interface implementations instead of a mix of package functions
and implementations.

- Add new unscopedEndpointsCounter
- Extract UpdateableReadyPodCounter from ReadyPodCounter
- Change autoscaler.New params to take UpdateableReadyPodCounter
- Make ReadyAddressCount() private
- Remove test of private readyAddressCount() method, this logic is
  covered by TestEndpointsAddressCounter

* Move private functions to bottom of file

* Rename endpoint counter creation function

NewEndpointAddressCounter -> NewScopedEndpointsCounter

Normally the old name and new name would be in the headline of
the commit, but they are too long to fit in the absurdly
narrow space provided by Github.

* Remove UpdateableReadyPodCounter type

In previous commits, the UpdateableReadyPodCounter type was introduced
to ensure that, upon autoscaler.Update(), the correct service name was
set on the UpdateableReadyPodCounter.

In conversation on PR knative/serving#4140, @vagabov pointed out that
the service name is intended to be immutable[0]. This means that
updating the service name is not a scenario that needs to be accounted
for.

[0] https://github.com/knative/serving/pull/4140#discussion_r286516170

* Rename name -> serviceName

* Rename NewUnscopedEndpointsCounter

NewUnscopedEndpointsCounter -> NewFixedEndpointsListCounter

Normally that would be the headline, but it's too long for
Github's narrow requirement.

The rename reflects the "unscoped" was meant to be symmetrical
with NewScopedEndpointsCounter, which takes an endpoints
informer, namespace and name as parameters.

But actually, this ReadyPodsCounter does not do an "unscoped"
version of the "scoped" ReadyPodsCounter -- that is, it does
not use an informer to retrieve all endpoints across all
namespaces. Instead it takes a predetermined list of endpoints
that are collected by the calling code.

* Add comments to exported funcs/types

* Backfill test for fixed endpoints counter

Coverage report showed that my changes had removed this path from
testing. This commit reintroduces a test focused on the narrower
behaviour.

- Rename TestEndpointAddressCounter->TestScopedEndpointsCounter

* Rename NewFixedEndpointsListCounter

to NewFixedEndpointsCounter

* Reintroduce resources.ReadyAddressCount()

This function is widely used in the codebase outside
the autoscaler, but not directly in the autoscaler
itself. In previous commits I created a PodReadyCounter
implementation using the same logic. When called across
the codebase, it never returned an error, leading to a
number of confusing ignored error returns.

- Remove unscopedCounter
- Remove NewFixedEndpointsCounter
- Rename resources.readyAddressCount -> resources.ReadyAddressCount
- Inline fetchReadyAddressCount into ReadyCount()