Inject ReadyPodCounter into NewServiceScraper (#4226)
PR knative/serving#4140 introduced the ReadyPodCounter type
and injected it into the autoscaler at creation time. This
made it possible to isolate the autoscaler's view of the
number of ready pods in simulations.

However, this left open the possibility that the stats
mechanism introduced in v0.6 could be out of sync with
the autoscaler, especially since there are key pieces of
logic in the statistics data path that may have an effect
on autoscaler behaviour.

This commit repeats the replacement of an EndpointsLister
with a ReadyPodCounter. This will make it possible for the
same counter to be injected into each of the statistics
and autoscaler decider components.