Factor calculation of final target concurrency out of the KPA. (#4249)
* Factor calculation of final target concurrency out of the KPA.

The KPA has some intricate calculations going on to figure out the final target concurrency based on defaults, the PA and the current config. This factors this knowledge out into a function that can be used both by the KPA and the HPA for consistency. It also centralizes this knowledge in one place (it used to be spread across the config and the PA).

* Review comments.

* Resolve package local conflict.