Add internal URL fields to the API coverage ignore list (#4271)
Knative Serving changed to use the internal knative/pkg/apis URL library
for parsing and marshaling the various URL fields such as `status.URL`
and `traffic[].URL` instead of manually manipulating the string type.
This new type in pkg uses the golang net/url library. This dependency
causes additional fields to be identified by the webhook coverage test.
As the API only interacts with URL as a string and the reconciler only
interacts with the Scheme and Host the other fields can be ignored.

See https://github.com/knative/pkg/blob/master/apis/url.go and
https://github.com/knative/serving/blob/master/pkg/reconciler/route/domains/domains.go#L79