Allow users to specify their container name.
With this change, users can specify the `name:` of the container they specify
in their revision spec.  This is to ease migration from K8s abstractions of
a similar shape where this field is required, and to lessen the eyesore of the
yaml returned by the API server, which doesn't `omitempty`.

The default value for container name is configurable via `config-defaults`, but
defaults to `user-container` for consistency with what we have today.  The
container name default in `config-defaults` is a Go template, which has access
to the ObjectMeta of the enclosing resource's ObjectMeta (e.g. Service,
Configuration), so if an operator wanted to make the container name match the
enclosing Service, they may set this to `{{.Name}}`.

Fixes: https://github.com/knative/serving/issues/4257